+++ CHANGELOG.md
@@ -4,9 +4,18 @@ NOWnews-api change logs
 
 ## [Unreleased]
 ### Added
-- 增加廣告版位在 `/promote/home` 端點 @esbb48
+- 新增兩個 tool 整理舊 Web AppInfo token 資料，請依序執行 @esbb48
+  - tools/cleanRepeatTokenForWeb.js
+  - tools/cleanInvalidToken.js
+- 新增 cronjob cleanInvalidToken @esbb48
 
+### Changed
+- 增加廣告版位在 `/promote/home` 端點 @esbb48
 - 軍聞社匯入新聞改分類 @appleoxxo
+- 區分 firebase prod 和 dev 的環境 @esbb48
+- 調整更新 POST `/app/info` 端點判斷重複的邏輯 @esbb48
+- 重構現有的 notification @esbb48
+- 避免重複的 token 被撈出 @esbb48
 
 ## 1.0.39 - 2017-10-17
 ### Changed